Implement Swift dataset copy
Align .NET with the structure of the resulting folder so that copy_dataset can be shared
