+++ environment/aws/topology_setup/test_server_platforms/swift_register.py
@@ -36,6 +36,7 @@ from environment.aws.topology_setup.test_server import (
     DOWNLOADED_TEST_SERVER_DIR,
     TEST_SERVER_DIR,
     TestServer,
+    copy_dataset,
 )
 
 from .ios_bridge import iOSBridge
@@ -84,6 +85,13 @@ class SwiftTestServer(TestServer):
         unzip_directory(download_file, FRAMEWORKS_DIR)
         download_file.unlink()
 
+    def _copy_dataset(self) -> None:
+        dest_dir = SWIFT_TEST_SERVER_DIR / "Assets"
+        copy_dataset(
+            dest_dir,
+            self.dataset_version,
+        )
+
 
 @TestServer.register("swift_ios")
 class SwiftTestServer_iOS(SwiftTestServer):
@@ -121,6 +129,7 @@ class SwiftTestServer_iOS(SwiftTestServer):
         return f"couchbase-lite-ios/{version_parts[0]}/{version_parts[1]}/testserver_ios.zip"
 
     def build(self) -> None:
+        self._copy_dataset()
         self._download_cbl()
         header("Building")
         env = os.environ.copy()
